refactor: moved benchmark test to top level testing service
