+++ am-lib/build.gradle
@@ -77,13 +77,9 @@ dependencies {
     compile group: 'org.hibernate.validator', name: 'hibernate-validator', version: '6.0.16.Final'
     compile group: 'org.aspectj', name: 'aspectjrt', version: '1.9.2'
     compile group: 'com.google.guava', name: 'guava', version: '27.1-jre'
-    compile group: 'org.springframework.boot', name: 'spring-boot-starter-jdbc', version: '2.1.3.RELEASE'
 
     testAnnotationProcessor group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
     testCompileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
-
-    annotationProcessor group: 'org.openjdk.jmh', name: 'jmh-generator-annprocess', version: '1.21'
-    compile group: 'org.openjdk.jmh', name: 'jmh-core', version: '1.21'
 
     testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter', version: '5.4.1'
     testCompile group: 'org.mockito', name: 'mockito-core', version: '2.25.1'
